requirements.txt
# Core LangChain dependencies
langchain
langchain-community
langchain-openai
langchain-huggingface
langchain-google-genai

# Vector database and embeddings
chromadb
sentence-transformers
tiktoken

# Data processing
pandas
numpy

# Environment and configuration
python-dotenv

# Streamlit web interface
streamlit

# Text processing and search
rank-bm25

# Utilities
tqdm
typing-extensions
